--- portal/.github/workflows/main.yml
@@ -1,43 +1,0 @@
-name: Standard Angular App Flow
-
-on:
-  pull_request:
-    branches: [ "main", "development" ]
-
-jobs:
-  lint:
-    name: Lint
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18'  # Adjust this to the Node.js version your project needs
-
-      - name: Install dependencies
-        run: npm install
-
-      - name: Run linting
-        run: npm run lint
-
-  build:
-    name: Build
-    runs-on: ubuntu-latest
-    needs: lint
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18'  # Adjust this to the Node.js version your project needs
-
-      - name: Install dependencies
-        run: npm install
-
-      - name: Build app
-        run: npm run build
