--- portal/.prettierrc.json
@@ -1,8 +1,0 @@
-{
-  "trailingComma": "all",
-  "tabWidth": 2,
-  "semi": true,
-  "singleQuote": true,
-  "bracketSpacing": true,
-  "printWidth": 100
-}
