--- portal/eslint.config.mjs
@@ -1,8 +1,0 @@
-import globals from "globals";
-import pluginJs from "@eslint/js";
-
-
-export default [
-  {languageOptions: { globals: globals.browser }},
-  pluginJs.configs.recommended,
-];
