--- portal/tsconfig.spec.json
@@ -1,9 +1,0 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./out-tsc/spec",
-    "types": ["jasmine"]
-  },
-  "include": ["src/**/*.spec.ts", "src/**/*.d.ts"]
-}
